end push swap with 800 for 100

+++ push_swap/init.sh
@@ -1,2 +1,2 @@
 
-gcc main.c && ./a.out > output.sh && bash output.sh && cat output.sh && rm output.sh
+clear && gcc main.c && ./a.out > output.sh && bash output.sh && cat output.sh && rm output.sh
